+++ src/main/java/me/borawski/hcf/command/commands/SetEndCommand.java
@@ -1,4 +1,4 @@
-package me.finestdev.components.commands;
+package me.borawski.hcf.command.commands;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
@@ -6,7 +6,7 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
 import me.borawski.hcf.Core;
-import me.finestdev.components.utils.Utils;
+import me.borawski.hcf.util.Utils;
 
 public class SetEndCommand implements CommandExecutor {
