+++ src/main/java/me/borawski/hcf/handler/EnderpearlHandler.java
@@ -1,4 +1,4 @@
-package me.finestdev.components.handlers;
+package me.borawski.hcf.handler;
 
 import java.util.Map;
 import java.util.UUID;
@@ -12,12 +12,12 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.block.Action;
 import org.bukkit.event.player.PlayerInteractEvent;
 
+import me.borawski.hcf.Components;
 import me.borawski.hcf.Core;
-import me.finestdev.components.Components;
-import me.finestdev.components.utils.Cooldown;
-import me.finestdev.components.utils.Cooldown.CooldownBase;
-import me.finestdev.components.utils.Cooldown.Time;
-import me.finestdev.components.utils.Utils;
+import me.borawski.hcf.util.Cooldown;
+import me.borawski.hcf.util.Utils;
+import me.borawski.hcf.util.Cooldown.CooldownBase;
+import me.borawski.hcf.util.Cooldown.Time;
 
 public class EnderpearlHandler implements Listener {
