+++ src/main/java/me/borawski/hcf/handler/MobStackHandler.java
@@ -1,4 +1,4 @@
-package me.finestdev.components.handlers;
+package me.borawski.hcf.handler;
 
 import java.util.ArrayList;
 import java.util.Iterator;
